refactor(llm/MyLedger): separate & rename MyLedger screens directories
