+++ apps/ledger-live-mobile/src/screens/MyLedgerChooseDevice/index.tsx
