+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/AppsList/AppFilter.tsx
