+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/AppsList/AppIcon.tsx
