+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/AppsList/AppInstallButton.tsx
