+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/AppsList/AppProgressButton.tsx
