+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/AppsList/AppRow.tsx
