+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/AppsList/AppStateButton.tsx
