+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/AppsList/AppUninstallButton.tsx
