+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/AppsList/AppUpdateAll.tsx
