+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/AppsList/AppUpdateButton.tsx
