+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/AppsList/AppUpdateStepper.tsx
