+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/AppsList/Searchbar.tsx
