+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/AppsScreen.tsx
