+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Benchmarking.tsx
