+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Device/CustomLockScreen.tsx
