+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Device/DeviceAppStorage.tsx
