+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Device/DeviceLanguage.tsx
