+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Device/DeviceLanguageSelection.tsx
