+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Device/DeviceName.tsx
