+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Device/DeviceOptionRow.tsx
