+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Device/StorageBarItem.tsx
