+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Device/images/dark/blue.png
