+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Device/images/dark/nanoS.png
