+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Device/images/dark/nanoX.png
