+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Device/images/light/blue.png
