+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Device/images/light/nanoX.png
