+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Device/index.tsx
