+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Modals/InstallAppDependenciesModal.tsx
