+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Modals/InstalledAppModal.tsx
