+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Modals/InstalledAppsModal.tsx
