+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Modals/QuitManagerModal.tsx
