+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Modals/StorageWarningModal.tsx
