+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/Modals/UninstallAppDependenciesModal.tsx
