+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/NotifBadge.tsx
