+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/index.tsx
