+++ apps/ledger-live-mobile/src/screens/MyLedgerDevice/shared.ts
